Project Setup Added Linters Restructured application folder and file structure Removed un-used and unwanted files Add svg files
